requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-apache-kafka>=1.9.0
apache-airflow-providers-smtp
apache-airflow-providers-openai
apache-airflow-providers-http
apache-airflow-providers-common-sql
apache-airflow-providers-common-messaging>=1.0.3